appengine/standard/flask/hello_world/requirements.txt
Flask==0.12
requests
requests-toolbelt
google-auth
